+++ src/main/java/com/sdl/selenium/extjs3/tab/TabPanel.java
@@ -81,17 +81,17 @@ public class TabPanel extends ExtJsComponent {
     public boolean setActive() {
         String baseTabPath = "//*[" + getPathBuilder().getBasePath() + "]";
         String titlePath = baseTabPath + getTitlePath();
-        WebLocator titleElement = new WebLocator(getContainer()).setElPath(titlePath).setInfoMessage(getText() + " Tab");
+        WebLocator titleElement = new WebLocator(getPathBuilder().getContainer()).setElPath(titlePath).setInfoMessage(getPathBuilder().getText() + " Tab");
         LOGGER.info("setActive : " + toString());
         boolean activated;
         try {
             activated = titleElement.click();
         } catch (ElementNotVisibleException e) {
             LOGGER.error("setActive Exception: " + e.getMessage());
-            WebLocator tabElement = new WebLocator(getContainer()).setElPath(baseTabPath);
+            WebLocator tabElement = new WebLocator(getPathBuilder().getContainer()).setElPath(baseTabPath);
             String id = tabElement.getAttributeId();
             String path = "//*[@id='" + id + "']//*[contains(@class, 'x-tab-strip-inner')]";
-            String script = "return Ext.getCmp('" + id + "').setActiveTab(" + getTabCount(getText(), path) + ");";
+            String script = "return Ext.getCmp('" + id + "').setActiveTab(" + getTabCount(getPathBuilder().getText(), path) + ");";
             LOGGER.warn("force TabPanel setActive with js: " + script);
             WebLocatorUtils.doExecuteScript(script);
             activated = true; // TODO verify when is not executed