+++ src/main/java/com/sdl/selenium/web/WebLocatorAbstractBuilder.java
@@ -30,7 +30,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setRoot(String)}
      * <p>root </p>
      * <pre>default to "//"</pre>
@@ -52,7 +52,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setTag(String)}
      * <p>tag (type of DOM element)</p>
      * <pre>default to "*"</pre>
@@ -74,7 +74,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setId(String)}
      */
     public String getId() {
@@ -94,7 +94,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setElPath(String)}
      * <p>returned value does not include containers path</p>
      */
@@ -116,7 +116,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setBaseCls(String)}
      */
     public String getBaseCls() {
@@ -136,7 +136,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setCls(String)}
      */
     public String getCls() {
@@ -161,7 +161,7 @@ public abstract class WebLocatorAbstractBuilder {
      * <pre>
      *     WebLocator element = new WebLocator().setClasses("bg-btn", "new-btn");
      * </pre>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setClasses(String...)}
      */
     public List<String> getClasses() {
@@ -184,7 +184,7 @@ public abstract class WebLocatorAbstractBuilder {
     }
 
     /**
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return WebLocators
      */
     public List<WebLocator> getChildNodes() {
@@ -198,7 +198,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setExcludeClasses(String...)}
      */
     public List<String> getExcludeClasses() {
@@ -218,7 +218,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setName(String)}
      */
     public String getName() {
@@ -238,7 +238,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setText(String, SearchType...)}
      */
     public String getText() {
@@ -259,7 +259,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setSearchTextType(SearchType...)}
      */
     public Set<SearchType> getSearchTextType() {
@@ -290,7 +290,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setStyle(String)}
      */
     public String getStyle() {
@@ -311,7 +311,7 @@ public abstract class WebLocatorAbstractBuilder {
     /**
      * <p><b>not implemented yet</b></p>
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setElCssSelector(String)}
      */
     public String getElCssSelector() {
@@ -333,7 +333,7 @@ public abstract class WebLocatorAbstractBuilder {
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
      * <p><b>Title only applies to Panel, and if you set the item "setTemplate("title", "text()='%s'")" a template.<b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setTitle(String)}
      */
     public String getTitle() {
@@ -410,7 +410,7 @@ public abstract class WebLocatorAbstractBuilder {
     }
 
     /**
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @param key
      * @return template
      */
@@ -420,7 +420,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used in logging process</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setInfoMessage(String)}
      */
     public String getInfoMessage() {
@@ -440,7 +440,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setVisibility(boolean)}
      */
     public boolean isVisibility() {
@@ -471,7 +471,7 @@ public abstract class WebLocatorAbstractBuilder {
     }
 
     /**
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return containter
      */
     public WebLocator getContainer() {
@@ -491,7 +491,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setLabel(String, SearchType...)}
      */
     public String getLabel() {
@@ -511,7 +511,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setLabel(String, SearchType...)}
      */
     public String getLabelTag() {
@@ -531,7 +531,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setLabelPosition(String)}
      */
     public String getLabelPosition() {
@@ -552,7 +552,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setPosition(int)}
      */
     public int getPosition() {
@@ -576,7 +576,7 @@ public abstract class WebLocatorAbstractBuilder {
 
     /**
      * <p><b><i>Used for finding element process (to generate xpath address)</i><b></p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return value that has been set in {@link #setType(String)}
      */
     public String getType() {
@@ -605,7 +605,7 @@ public abstract class WebLocatorAbstractBuilder {
     /**
      * <p>Used only to identify class type of current object</p>
      * <p> Not used for css class!</p>
-     * @deprecated use getPathBuilder().get*
+     * @deprecated use {@link #getPathBuilder()}.get to get path attributes
      * @return string
      */
     public String getClassName() {